+++ CHANGELOG.md
@@ -11,6 +11,11 @@ Types of changes:
 - `Removed` for now removed features.
 - `Fixed` for any bug fixes.
 
+## [Unreleased]
+
+### Added
+- Added SNV read position metrics JSON and plot (#360)
+
 ## [RC5.0.0] - 2026-03-13
 
 ### Added